Created a basic Face API test app
To test the validity of the Microsoft Face API, the app takes a photo
and outputs the emotions to the log. In order to run the app, add the
Microsoft Face API the user-specific `build.properties` file as
`FreqOut_FaceAPIKey`. This is just a large test and will not be merged
to master.